requirements.txt
aniso8601==9.0.1
attrs==21.2.0
autopep8==1.6.0
certifi==2021.10.8
charset-normalizer==2.0.6
click==8.0.1
dominate==2.6.0
Flask==2.0.1
Flask-Bootstrap==3.3.7.1
Flask-Cors==3.0.10
Flask-Login==0.5.0
Flask-RESTful==0.3.9
Flask-SQLAlchemy==2.5.1
greenlet==1.1.1
idna==3.2
importlib-metadata==4.8.1
iniconfig==1.1.1
itsdangerous==2.0.1
Jinja2==3.0.1
MarkupSafe==2.0.1
packaging==21.2
pkg-resources==0.0.0
pluggy==1.0.0
py==1.11.0
pycodestyle==2.8.0
PyMySQL==1.0.2
pyparsing==2.4.7
pytest==6.2.5
python-dotenv==0.19.0
pytz==2021.1
requests==2.26.0
requests-mock==1.9.3
six==1.16.0
SQLAlchemy==1.4.23
toml==0.10.2
typing-extensions==3.10.0.2
urllib3==1.26.7
visitor==0.1.3
Werkzeug==2.0.1
zipp==3.6.0
